requirements.txt
anyio==3.6.2
APScheduler==3.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
charset-normalizer==3.0.1
click==8.1.3
DateTime==5.0
dnspython==2.3.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
llvmlite==0.39.1
mysql-connector-python==8.0.32
numba==0.56.4
numpy==1.23.5
pandas==1.5.3
protobuf==3.20.3
pybithumb==1.0.21
pydantic==1.10.4
pymongo==4.3.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
starlette==0.22.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
zope.interface==5.5.2
